split hostentry into its own file
