+++ sia/hostdb/hostdb.go
@@ -3,7 +3,6 @@ package hostdb
 import (
 	// "crypto/rand"
 	// "errors"
-	"math"
 	// "math/big"
 	"sync"
 
@@ -41,49 +40,12 @@ type HostAnnouncement struct {
 	FreezeIndex     uint64 // The index of the output that froze coins.
 }
 
-// the Host struct is kept in the client package because it's what the client
-// uses to weigh hosts and pick them out when storing files.
-type HostEntry struct {
-	IPAddress   network.Address
-	MinFilesize uint64
-	MaxFilesize uint64
-	MinDuration consensus.BlockHeight
-	MaxDuration consensus.BlockHeight
-	Window      consensus.BlockHeight
-	Tolerance   uint64
-	Price       consensus.Currency
-	Burn        consensus.Currency
-	Freeze      consensus.Currency
-	CoinAddress consensus.CoinAddress
-}
-
+// New returns an empty HostDatabase.
 func New() (hdb *HostDatabase) {
-	hdb = new(HostDatabase)
+	hdb = &HostDatabase{
+		inactiveHosts: make(map[string]*hostNode),
+	}
 	return
-}
-
-// host.Weight() determines the weight of a specific host, which is:
-//
-//		Freeze * Burn / square(Price).
-//
-// Freeze has to be linear, because any non-linear freeze will invite sybil
-// attacks.
-//
-// For now, Burn is also linear because an increased burn means increased risk
-// for the host (Freeze on the other hand has no risk). It might be better to
-// make burn grow sublinearly, such as taking sqrt(Burn) or burn^(4/5).
-//
-// We take the square of the price to heavily emphasize hosts that have a low
-// price. This is also a bit simplistic however, because we're not sure what
-// the host might be charging for bandwidth.
-func (h *HostEntry) Weight() consensus.Currency {
-	// adjustedBurn := math.Sqrt(float64(h.Burn))
-	adjustedBurn := float64(h.Burn)
-	adjustedFreeze := float64(h.Freeze)
-	adjustedPrice := math.Sqrt(float64(h.Price))
-
-	weight := adjustedFreeze * adjustedBurn / adjustedPrice
-	return consensus.Currency(weight)
 }
 
 /*
